+++ src/controllers/index.ts
@@ -48,6 +48,7 @@ export abstract class ControllerBase<TApp extends any = IHttpServer> implements 
 
 export * from "./Authorize";
 export * from "./Body";
+export * from "./BodyParseErrorHandler";
 export * from "./CONNECT";
 export * from "./Controller";
 export * from "./DELETE";